prettify ProgressBar.cs + move usefull scripts to Scripts folder
