--- Assets/CamPosition.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f555440c08ded1840b1d473aa3cc0a2a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
