--- Assets/Prefabs/progressbar/Health.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 5203f8c3f8ecad84e9349a3f32b4e90c
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
