+++ Assets/Scripts/ClearLevel.cs
