+++ Assets/Scripts/InfoPanelManager.cs
