+++ Assets/Scripts/InfoPanelManager.cs.meta
