+++ Assets/Scripts/MiniCollider.cs
@@ -1,7 +1,8 @@
-﻿using System.Collections;
+using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.UI;
+
 public class MiniCollider : MonoBehaviour
 {
 	Target parent;
