--- Assets/Vanya's_Files/Health.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 975941546686224468b259a3fa7b9f23
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
