--- Assets/Vanya's_Files/Readme.asset
@@ -1,64 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: fcf7219bab7fe46a1ad266029b2fee19, type: 3}
-  m_Name: Readme
-  m_EditorClassIdentifier: 
-  icon: {fileID: 2800000, guid: 7801804018a7dcf42abb827444e18660, type: 3}
-  title: Universal Render Pipeline Template
-  sections:
-  - heading: Universal Render Pipeline
-    text: The Universal Project Template configures Project settings for Projects
-      where performance, wide platform support, and ease of customizing graphics
-      are the primary considerations.
-    linkText: 
-    url: 
-  - heading: 
-    text: This Template uses the Universal Render Pipeline (URP) and Shader Graph.
-    linkText: 
-    url: 
-  - heading: 
-    text: URP is prebuilt Scriptable Render Pipeline that is quick and easy to customize,
-      and lets you create optimized graphics across a wide range of platforms. URP
-      also includes an optimized 2D renderer complete with 2D lights and pixel perfect
-      rendering, and an integrated post-processing solution.
-    linkText: 
-    url: 
-  - heading: 
-    text: Shader Graph is a tool that allows you to create shaders using a visual
-      node editor instead of writing code.
-    linkText: 
-    url: 
-  - heading: 
-    text: This template contains a sample Scene that contains examples of how to
-      configure lighting settings, Materials, Shaders, and post-processing effects
-      in URP, several preconfigured Universal Render Pipeline Assets that let you
-      quickly swap between graphics quality levels, and Presets that have been optimized
-      for use with URP.
-    linkText: 
-    url: 
-  - heading: 
-    text: This template contains a sample Scene that contains examples of how to
-      configure lighting settings, Materials, Shaders, and post-processing effects
-      in URP, several preconfigured Universal Render Pipeline Assets that let you
-      quickly swap between graphics quality levels, and Presets that have been optimized
-      for use with URP.
-    linkText: 
-    url: 
-  - heading: 
-    text: 'To read more about URP and its built-in features, see the '
-    linkText: URP documentation.
-    url: https://docs.unity3d.com/Packages/com.unity.render-pipelines.universal@latest/index.html
-  - heading: 
-    text: 'For more information about Shader Graph, see the '
-    linkText: Shader Graph documentation
-    url: https://docs.unity3d.com/Packages/com.unity.shadergraph@latest
-  loadedLayout: 1
