--- Assets/Vanya's_Files/Readme.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 83c2ed844a8c74b779a4c823d16594b1
-timeCreated: 1484217493
-licenseType: Store
-NativeFormatImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
