--- Assets/Vanya's_Files/progressbar.rar.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 3a9572fe83d361245b94af48200707a6
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
